Issue #10908: Moved tests from profile to os_pages
The existing tests are related to os_pages.
